Change `force-file-write` default to be `true`.
Fixes #1000

This is a usability issue. Almost everyone will expect mockery to simply overwrite files. My original
thinking was that we want to protect mockery from overwriting user files that might be unrelated to
mockery, but version control is a thing and we should err on the side of making mockery more intuitive.