Fix variadic parameter handling in `.EXPECT().Run()`
An issue was reported where the `.Run()` method of testify mocks
was not handling variadic parameters correctly when `unroll-variadic: False`.
It was attempting to grab the variadic parameters one by one by iterating
through the argument list, when instead it should have just grabbed the
last argument and type asserted it as a slice of the variadic type.

Fixes #1024