Disable `goimports` from messing with imports
It might sound odd to not allow goimports to add/delete imports, but this was causing problems in #1019. As a matter of principle, we should not be relying on a tool to fix imports for us. If imports are missing or redundant, that is a problem with the template (or mockery) that should be fixed.